Fix active/stable toggling when all checkboxes are unchecked
Checkbox inputs don't pass values through to the controller when
the boxes are unchecked. For the active and stable labels, the
controller needs to treat the value not being there as being unchecked,
and that includes if the entire dictionary of that label is not included,
as that means all labels are unchecked.